+++ packages/@react-native-windows/automation-channel/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @react-native-windows/automation-channel
 
-<!-- This log was last generated on Tue, 27 Jan 2026 19:34:00 GMT and should not be manually modified. -->
+<!-- This log was last generated on Wed, 11 Feb 2026 16:10:55 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 0.81.4
+
+Wed, 11 Feb 2026 16:10:55 GMT
+
+### Patches
+
+- Bump react-native-windows to v0.81.4
 
 ## 0.81.3